+++ ejo-base.php
@@ -159,6 +159,9 @@ final class EJO_Base
         /* Shortcodes */
         require_once( self::$dir . 'includes/base/shortcodes/shortcodes.php' );  
 
+        /* Footer Line Widget */
+        require_once( self::$dir . 'includes/base/footer-line-widget/footer-line-widget.php');
+
         /* Theme Functions */
         require_once( self::$dir . 'includes/base/theme-functions/post-summary.php' );  
         require_once( self::$dir . 'includes/base/theme-functions/sidebar-widget-count.php' );