+++ CHANGES.md
@@ -1,3 +1,4 @@
+- Added Footer Line Widget
 - Added helper function ejo_get_image_dimensions
 - Fixed secondary custom color bug